refactor: break inputiino into multiple namespaces
