+++ cmake/compile_definitions/linux.cmake
@@ -223,11 +223,23 @@ if(${SUNSHINE_ENABLE_TRAY} AND ${SUNSHINE_TRAY} EQUAL 0 AND SUNSHINE_REQUIRE_TRA
 endif()
 
 if(${SUNSHINE_USE_LEGACY_INPUT})
-    list(APPEND PLATFORM_TARGET_FILES "${CMAKE_SOURCE_DIR}/src/platform/linux/input.cpp")
+    list(APPEND PLATFORM_TARGET_FILES "${CMAKE_SOURCE_DIR}/src/platform/linux/input/legacy_input.cpp")
 else()
     add_subdirectory("${CMAKE_SOURCE_DIR}/third-party/inputtino/")
     list(APPEND SUNSHINE_EXTERNAL_LIBRARIES inputtino::libinputtino)
-    list(APPEND PLATFORM_TARGET_FILES "${CMAKE_SOURCE_DIR}/src/platform/linux/inputtino.cpp")
+    list(APPEND PLATFORM_TARGET_FILES
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_common.h"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_gamepad.h"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_gamepad.cpp"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_keyboard.h"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_keyboard.cpp"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_mouse.h"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_mouse.cpp"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_pen.h"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_pen.cpp"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_touch.h"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino_touch.cpp"
+            "${CMAKE_SOURCE_DIR}/src/platform/linux/input/inputtino.cpp")
 endif()
 
 list(APPEND PLATFORM_TARGET_FILES
