+++ src/platform/linux/input/inputtino_mouse.h
@@ -1,0 +1,28 @@
+#pragma once
+
+#include <boost/locale.hpp>
+#include <inputtino/input.hpp>
+#include <libevdev/libevdev.h>
+
+#include "src/platform/common.h"
+
+#include "inputtino_common.h"
+
+using namespace std::literals;
+
+namespace platf::mouse {
+  void
+  move(input_raw_t *raw, int deltaX, int deltaY);
+
+  void
+  move_abs(input_raw_t *raw, const touch_port_t &touch_port, float x, float y);
+
+  void
+  button(input_raw_t *raw, int button, bool release);
+
+  void
+  scroll(input_raw_t *raw, int high_res_distance);
+
+  void
+  hscroll(input_raw_t *raw, int high_res_distance);
+}  // namespace platf::mouse
